Changed project to deal data from MySQL using JPA

+++ src/main/java/com/CofeeShop/CoffeeShopMenuApplication.java
@@ -10,16 +10,13 @@ import com.CofeeShop.model.Product;
 public class CoffeeShopMenuApplication {
     public static void main(String[] args) {
         ApplicationContext context = SpringApplication.run(CoffeeShopMenuApplication.class, args);
-
+        
         // Product instantiation and printing details
-        Product espresso = new Product(1, "Espresso", 2.99);
-        Product croissant = new Product(2, "Croissant", 3.49);
+        Product espresso = new Product((long)1,"Espresso", 2.99);
+        Product croissant = new Product((long)2,"Croissant", 3.49);
     
+        
         System.out.println("Product ID: " + espresso.getId() + " Product Name: " + espresso.getName() + " Product Price: " + espresso.getPrice());
         System.out.println("Product ID: " + croissant.getId() + " Product Name: " + croissant.getName() + " Product Price: " + croissant.getPrice());
-        
-        // Clean shutdown
-//        int exitCode = SpringApplication.exit(context, () -> 0);
-  //      System.exit(exitCode);
     }
 }
